task-engine/downsample/requirements.txt
influxdb-client
flightsql-dbapi
pandas